baza de date 2

+++ src/models/administration/Administration.java
@@ -6,11 +6,11 @@ public class Administration {
     private Integer id;
     private Prescription prescription;
     private Medication medication;
-    private Float quantity;
-    private Float frequency;
+    private Integer quantity;
+    private Integer frequency;
     public Administration() {}
     public Administration(Integer id, Prescription prescription, Medication medication,
-                          Float quantity, Float frequency)
+                          Integer quantity, Integer frequency)
     {
         this.id = id;
         this.prescription = prescription;
@@ -19,7 +19,7 @@ public class Administration {
         this.frequency = frequency;
     }
     public Administration(Prescription prescription, Medication medication,
-                          Float quantity, Float frequency)
+                          Integer quantity, Integer frequency)
     {
         this.prescription = prescription;
         this.medication = medication;
@@ -51,19 +51,19 @@ public class Administration {
         this.medication = medication;
     }
 
-    public Float getQuantity() {
+    public Integer getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Float quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
 
-    public Float getFrequency() {
+    public Integer getFrequency() {
         return frequency;
     }
 
-    public void setFrequency(Float frequency) {
+    public void setFrequency(Integer frequency) {
         this.frequency = frequency;
     }
 }
